Check for valid id and return not found if not valid/existing.
Comply with standard REST response protocol.

+++ src/form_collection.py
@@ -1,5 +1,13 @@
 from pymongo import MongoClient
 from bson.objectid import ObjectId
+from bson.errors import InvalidId
+
+def is_valid_object_id(obj_id):
+    try:
+        wid = ObjectId(obj_id)
+        return True
+    except InvalidId:
+        return False
 
 
 access_url = 'forms-editor-datastore:27017'
@@ -13,9 +21,11 @@ class FormCollection:
         self.form_collection = self.db.form_collection
 
     def get_one_form(self, fid):
+        if not is_valid_object_id(fid):
+            raise ValueError
         form = self.form_collection.find_one(ObjectId(fid))
         if not form:
-            return 'No results'
+            raise ValueError
         form['_id'] = str(form['_id'])
         return form
 
@@ -32,12 +42,19 @@ class FormCollection:
         return str(fid)
 
     def update_one_form(self, fid, updates):
+        if not is_valid_object_id(fid):
+            raise ValueError
         updates = {'$set': updates}
-        self.form_collection.update_one({'_id': ObjectId(fid)}, updates)
+        update_res = self.form_collection.update_one({'_id': ObjectId(fid)}, updates)
+        if update_res.matched_count == 0:
+            raise ValueError
 
     def delete_one_form(self, fid):
-        self.form_collection.delete_one({'_id': ObjectId(fid)})
-        return True
+        if not is_valid_object_id(fid):
+            raise ValueError
+        del_res = self.form_collection.delete_one({'_id': ObjectId(fid)})
+        if del_res.deleted_count == 0:
+            raise ValueError
 
     def delete_all_forms(self):
         self.form_collection.delete_many({})
